+++ src/Render/Entity_DrawSetup.cpp
@@ -174,7 +174,7 @@ void __cdecl FUN_00504b50(int param_1, int param_2, float *param_3,
         *(char *)((int)this_ + 0x88) = 0;
         FUN_00441e00(this_, 2, *(float *)(param_1 + 0x168), *(int *)(param_1 + 100),
                      *(float *)(param_1 + 0x68), *(float *)(param_1 + 0x6c),
-                     *(float *)(param_1 + 0x70), *(float *)(param_1 + 0x58), 1170);
+                     *(float *)(param_1 + 0x70), *(int *)(param_1 + 0x58), 1170);
         *(char *)((int)this_ + 0x88) = -1;
     }
     else if (effectType == 831) {
@@ -200,10 +200,10 @@ void __cdecl FUN_00504b50(int param_1, int param_2, float *param_3,
         }
         FUN_00441e00(this_, 8, *(float *)(param_1 + 0x168), *(int *)(param_1 + 100),
                      *(float *)(param_1 + 0x68), *(float *)(param_1 + 0x6c),
-                     *(float *)(param_1 + 0x70), *(float *)(param_1 + 0x58), 1171);
+                     *(float *)(param_1 + 0x70), *(int *)(param_1 + 0x58), 1171);
         FUN_00441e00(this_, 0x44, *(float *)(param_1 + 0x168), *(int *)(param_1 + 100),
                      *(float *)(param_1 + 0x68), *(float *)(param_1 + 0x6c),
-                     *(float *)(param_1 + 0x70), *(float *)(param_1 + 0x58), 1171);
+                     *(float *)(param_1 + 0x70), *(int *)(param_1 + 0x58), 1171);
         return;
     }
     else if (effectType == 958) {
@@ -212,10 +212,10 @@ void __cdecl FUN_00504b50(int param_1, int param_2, float *param_3,
         *(float *)((int)this_ + 0x50) = 1.0f;
         FUN_00441e00(this_, 2, *(float *)(param_1 + 0x168), *(int *)(param_1 + 100),
                      *(float *)(param_1 + 0x68), *(float *)(param_1 + 0x6c),
-                     *(float *)(param_1 + 0x70), -1.0f, -1);
+                     *(float *)(param_1 + 0x70), -1, -1);
         FUN_00441e00(this_, 0x44, 0.5f, *(int *)(param_1 + 100),
                      *(float *)(param_1 + 0x68), *(float *)(param_1 + 0x6c),
-                     *(float *)(param_1 + 0x70), -1.0f, 1171);
+                     *(float *)(param_1 + 0x70), -1, 1171);
         return;
     }
 
@@ -231,14 +231,14 @@ void __cdecl FUN_00504b50(int param_1, int param_2, float *param_3,
         *(char *)(DAT_05828d58 + effectType * 0xbc + 0x88) = 0;
         FUN_00441e00(this_, 2, *(float *)(param_1 + 0x168), *(int *)(param_1 + 100),
                      *(float *)(param_1 + 0x68), *(float *)(param_1 + 0x6c),
-                     *(float *)(param_1 + 0x70), *(float *)(param_1 + 0x58), 1170);
+                     *(float *)(param_1 + 0x70), *(int *)(param_1 + 0x58), 1170);
         *(char *)(DAT_05828d58 + effectType * 0xbc + 0x88) = -1;
         *(float *)((int)this_ + 0x48) = 1.0f;
         *(float *)((int)this_ + 0x4c) = 1.0f;
         *(float *)((int)this_ + 0x50) = 1.0f;
         FUN_00441e00(this_, 2, *(float *)(param_1 + 0x168), *(int *)(param_1 + 100),
                      *(float *)(param_1 + 0x68), *(float *)(param_1 + 0x6c),
-                     *(float *)(param_1 + 0x70), *(float *)(param_1 + 0x58), -1);
+                     *(float *)(param_1 + 0x70), *(int *)(param_1 + 0x58), -1);
         return;
     }
     else if (effectType == 834) {
@@ -249,7 +249,7 @@ void __cdecl FUN_00504b50(int param_1, int param_2, float *param_3,
         pulse = sinf((float)WorldTime * 0.002f) * 0.3f + 0.7f;
         FUN_00441e00(this_, 0x42, 1.0f, 0, pulse,
                      *(float *)(param_1 + 0x6c), *(float *)(param_1 + 0x70),
-                     *(float *)(param_1 + 0x58), -1);
+                     *(int *)(param_1 + 0x58), -1);
         return;
     }
     else if (effectType == 865) {
@@ -511,6 +511,6 @@ LAB_after_render: ;
                      *(float*)(param_1 + 0x68),
                      *(float*)(param_1 + 0x6c),
                      *(float*)(param_1 + 0x70),
-                     -1.0f, 0xffffffff);
+                     -1, 0xffffffff);
     }
 }
